Move `inplace_batch_inverse` to `FieldElement::*`
